Added Host Button functionality
Bound callback functions to our buttons and specifically added functionality to the Host Button. It now calls our subsystem's CreateSession method which calls the Interface's CreateSession method, effectively creating an online session. Whether that is successful or not, we are brought to the Lobby level. Also added the ability to press ESC to quit the game.