Created logic for adding a session
Created c++ logic for adding a session entry to the widget. Also learned that we MUST create the Blueprint Widget version, not the C++ class. This is because the widgets on the WBP aren't constructed if we create the class. For reference, we must create a WBP_SessionMenu and not a USessionMenu if we wish to access properties we have bound to widgets such as the List.