Created MenuHelper utility class
Created a class called MenuHelper that houses information regarding how the Maps and Game Modes dropdowns should be populated and methods for populating those dropdowns. This was so I didn't have to duplicate code across multiple widgets that display the list of maps and game modes. Also makes the project more scalable and configurable.