Implement Searching through Steam API
We are now searching for lobbies using the Steam API directly. However, there is the issue that our Blueprint Async nodes are not working properly in C++ and we cannot bind our delegates properly. I have to convert these async classes to a UObject to make this work now. That's tomorrow's project